fix(invitations): clean up prior Twilio conversations on re-send (#16)
Twilio refuses to bind the same (phone, proxy) pair to two active
Conversations simultaneously. That made Send SMS fail with
"Invalid messaging binding address" / "Participant already exists
in another conversation" the second time the bishop clicked Send
for the same speaker (or even after a first successful send, if
they retried at all).

sendSpeakerInvitation now calls cleanupPriorConversations() before
createConversation(): queries Firestore for speakerInvitations
matching (wardId, speakerId, meetingDate), pulls their
conversationSids, and deletes each via Twilio's API. Per-deletion
failures are logged and swallowed — one stale/missing SID can't
block the re-send path.

The Firestore docs stay — they're part of the audit trail.
useLatestInvitation still surfaces the newest one to the bishop's
Prepare page; older docs point at deleted SIDs which a future
cleanup job could reap, but that's out of scope for this commit.

File size — had to extract sendSpeakerInvitation.helpers.ts
(assertActiveMember, cleanupPriorConversations, tryAddSms,
tryEmail, trySms) to stay under the 150-LOC ceiling.
